Refactor imports and styles: update import paths for components, remove unused CardStyles file, and enhance structure for better organization.
