+++ src/Sections/MainBody/Updates/Updates.jsx
@@ -2,17 +2,22 @@ import { UpdateCard } from "../../../Components/Cards/UpdateCard";
 import {
   SectionContainer,
   CardContainer,
-} from "../../../Components/CardSectionStyling";
-import updates from "../../../data/Updates.json";
+} from "../../../Components/Styles/CardSectionStyling";
+import updates from "../../../Data/Updates.json";
+// Validate updates structure
+if (!Array.isArray(updates.updates)) {
+  console.error(
+    "Invalid updates structure: 'updates.updates' should be an array."
+  );
+}
 
 // Ensure updates has a valid structure
 
 export const Updates = () => {
   return (
     <SectionContainer id="updates">
-      <div className="h2-title">
-        <h2>Updates</h2>
-      </div>
+      <h2>Updates</h2>
+
       <CardContainer className="updates-container">
         {updates.updates.map((update, index) => (
           <div key={index} className="updates-card">
